chore: rename shortcut model filename
